Add fraudStatus, add transaction to company operations

+++ src/companies/companies-data.service.ts
@@ -1,4 +1,4 @@
-import { Injectable } from '@nestjs/common';
+import { Injectable, NotFoundException } from '@nestjs/common';
 import { PrismaService } from 'src/prisma/prisma.service';
 import { CreateCompanyDto } from './dto/create-company.dto';
 import { UpdateCompanyDto } from './dto/update-company.dto';
@@ -20,10 +20,49 @@ export class CompaniesDataService {
     });
   }
 
-  async delete(id: string) {
-    return this.prisma.company.delete({
+  async findOne(options: {
+    where: Record<string, any>;
+    include?: Record<string, boolean>;
+  }) {
+    return await this.prisma.company.findFirst({
+      where: options.where,
+      include: options.include,
+    });
+  }
+
+  async delete(id: string, tx?: PrismaService) {
+    if (tx) {
+      return await tx.company.delete({
+        where: { id },
+      });
+    }
+
+    return await this.prisma.company.delete({
       where: { id },
     });
+
+    //todo: Move acl and userCompany deletes in their data-service
+
+    // return this.prisma.$transaction(async (tx) => {
+    //   // вынести в свой датасервис
+    //   await tx.acl.deleteMany({
+    //     where: {
+    //       OR: [
+    //         { resource: `/companyManagement/${id}` },
+    //         { resource: `/company/${id}` },
+    //       ],
+    //     },
+    //   });
+
+    //   // вынести в свой датасервис ( тоже прокидываю необязательный tx)
+    //   await tx.userCompany.deleteMany({
+    //     where: { companyId: id },
+    //   });
+
+    //   return await tx.company.delete({
+    //     where: { id },
+    //   });
+    // });
   }
 
   async deletePendingCompanies() {
